Restructure MoL - work in progress
